Refactor request handling for improved security and structure
- Instead of redirecting requests to another page, handle data on the same page.
- Include shared files using the `includes/` directory.
- Removed access to the `app/` directory to enhance security.

+++ app/includes/thesards-mgr.inc.php
@@ -1,24 +1,21 @@
 <?php
 
-if ($_SERVER["REQUEST_METHOD"] === "GET" && isset($_GET["action"])) {
+if ($_SERVER["REQUEST_METHOD"] === "GET" && isset($_GET["action"]) && empty($_GET["error"])) {
     $action = $_GET["action"];
 
-    include_once "../classes/models/dbh.class.php";
-    include_once "../classes/models/thesards_mgr_model.class.php";
-    include_once "../classes/controllers/thesards_mgr_contr.class.php";
+    include_once __DIR__ . "/../classes/models/dbh.class.php";
+    include_once __DIR__ . "/../classes/models/thesards_mgr_model.class.php";
+    include_once __DIR__ . "/../classes/controllers/thesards_mgr_contr.class.php";
     
     if ($action === "delete") {
         $manager = new ThesardsMgrContr($_GET["id"]);
         $manager->delete();
-        header("location: ../../public/gestion-des-thesards.php?action=delete&error=none");
+        header("location: gestion-des-thesards.php?action=delete&error=none");
         exit();
     } elseif ($action === "add") {
         $manager = new ThesardsMgrContr($_GET["firstname"], $_GET["lastname"], $_GET["email"], $_GET["password"]);
         $manager->add();
-        header("location: ../../public/gestion-des-thesards.php?action=add&error=none");
+        header("location: gestion-des-thesards.php?action=add&error=none");
         exit();
     }
-} else {
-    header("location: ../../public/dashboard.php");
-    exit();
 }
